% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ihealth_distance}
\alias{ihealth_distance}
\title{Calculate distance traveled}
\usage{
ihealth_distance(
  data,
  start = NA,
  end = NA,
  method = c("max", "distance", "steps"),
  steps_per_mile = 2000
)
}
\arguments{
\item{data}{Output from ihealth_read}

\item{start}{First date to include.  If NA, the oldest records are used.}

\item{end}{Last date to include. If NA, the most recent records are used.}

\item{method}{Approach to summarize distance by foot.  If `distance`, then
miles are used.  If `steps`, then the number of steps is used to estimate distance.
If `max`, then the max of `distance` and `steps` is used.}

\item{steps_per_mile}{The conversion rate from steps to miles}
}
\value{
A data frame
}
\description{
Get distance traveled by foot and bike (walking, running, cycling). Currently,
only records recorded in miles are accepted.
}
