% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{ihealth_read}
\alias{ihealth_read}
\title{Parse Apple Health Export}
\usage{
ihealth_read(path, verbose = TRUE)
}
\arguments{
\item{path}{Path to Apple Health directory}

\item{verbose}{Print status messages}
}
\value{
A data frame
}
\description{
Read the exported XML data from Apple Health
}
